its unlikely someone would have 2^31-1 and more than 2^15-1 packages, and similarly, its extremely slim that someone would have half a package (eg 875.3), so change the u to an i and instead of i32 it should be an i16
which is just a fancy way of saying "less memory"

+++ src/packages.rs
@@ -3,7 +3,7 @@
 
 use std::process::{Command, Stdio}; // For executing shell commands.
 
-pub fn get_num_packages() -> u32 {
+pub fn get_num_packages() -> i16 {
     let num_packages = packages_generic("pacman", &["-Q"])
         .or_else(|_| packages_generic("yum", &["list", "installed"]))
         .or_else(|_| packages_generic("dpkg-query", &["-l"]))
@@ -17,7 +17,7 @@ pub fn get_num_packages() -> u32 {
         .unwrap_or_else(|_| "Unknown".to_string());
 
     // Count the total number of packages
-    num_packages.lines().count() as u32
+    num_packages.lines().count() as i16
 }
 
 pub fn packages_generic(cmd: &str, options: &[&str]) -> Result<String, String> {
